Refactor: exception 단일 패키지에 ResponseCode 보관
